requirements.txt
fastapi==0.111.*
uvicorn[standard]==0.30.*
python-docx==1.1.*
python-multipart==0.0.9
pydantic==2.*
mammoth==1.6.*

